fix: code-review P1+P2 — preflight, recorder lifecycle, silence detect, abs overflow, tap-toggle
Six issues from code review, fixed in one batch.

P1.1 — HUD bypassed MeetingPreflight.gate()
The floating Meeting panel's Record button called
MeetingTranscriptionSession.shared.start() directly. The only gate
call lived in the unused MenuBarController.startMeetingTranscript
path. Recording could begin with Meeting Mode off, no consent, or no
API key.

Fix: preflight gate runs inside start() itself with an injectable
provider (default calls MeetingPreflight.gate with live AppSettings;
tests pass `{ .success(()) }` to bypass). HUD's onStart now surfaces
the preflight error as an NSAlert instead of swallowing it.

P1.2 — Silent file reported as audible
SilenceTrim.detectBounds returned (firstAudibleSec: 0,
lastAudibleSec: totalDur) when no sample exceeded the RMS threshold,
so any silent file longer than 100ms passed `hasAudibleContent` and
was sent to Whisper, defeating the hallucination-prevention path.

Fix: AudibleBounds gains `hadAudibleSamples: Bool`. hasAudibleContent
requires it. Silent path returns hadAudibleSamples=false; trim()
correctly throws .noAudibleContent. New SilenceTrimTests covers the
silent and short-blip cases.

P1.3 — Failed start left session pinned at .recording
start() saved the session as .recording on disk, then `try await
system.start()` with no catch. A throw left memory + disk stuck
recording; subsequent start()s threw .alreadyActive forever (until
cold restart). This was the in-the-wild blocker reported during 0.6.5
testing.

Fix: wrap recorder.start() in do/catch that clears the in-memory
session and persists status=.failed with a reason. Surfaces a new
SessionError.recorderStartFailed.

P1.4 — Failed stop left session pinned at .chunking
Symmetric bug in stop(): session saved as .chunking then `try await
system.stop()` propagated unconditionally.

Fix: wrap recorder.stop() in do/catch. Mark session .failed, also
stop the mic recorder so it's not left dangling, clear in-memory
state, then rethrow.

P2.1 — Tap-toggle ignored Fn / single-modifier hotkeys
handleTapToggle filtered to .keyDown only, but the dispatcher routes
flagsChanged for .fn and .modifier keys. Selecting tap-to-toggle on
the default Fn binding silently disabled recording entirely.

Fix: handleTapToggle now accepts .flagsChanged for modifier keys and
toggles on the rising edge (last-match=false → now-match=true) using
a new tapModifierLastMatch state field — each physical tap fires
exactly once.

P2.2 — abs(Int16.min) trap in audio callback
First-buffer peak probe in AudioRecorder used abs(int16Channel[f])
on Int16. A clipped sample of -32768 traps because +32768 isn't
representable as Int16. Could crash the audio callback.

Fix: widen to Int before abs (`abs(Int(int16Channel[f]))`).

Tests: 7 new (3 SilenceTrim, 3 MeetingTranscriptionSession recovery
+ preflight, 1 HotkeyMonitor Fn-tap matching). 281 total green.

+++ Sources/vox/Audio/AudioRecorder.swift
@@ -188,9 +188,12 @@ public final class AudioRecorder {
         // can tell the difference between "tap not firing" (no log) vs "tap
         // firing with silent samples" (log shows peak=0).
         if pcmBytesWritten == 0 {
-            var peak: Int16 = 0
+            // Widen to Int before negating: `abs(Int16.min)` traps because
+            // +32768 is not representable as Int16 (a clipped sample of
+            // -32768 from a hot mic would crash the audio callback).
+            var peak: Int = 0
             for f in 0..<Int(outBuf.frameLength) {
-                let s = abs(int16Channel[f])
+                let s = abs(Int(int16Channel[f]))
                 if s > peak { peak = s }
             }
             dlog("AudioRecorder first buffer frames=\(outBuf.frameLength) peak=\(peak)")
